Move new functions/methods for easier unitesting
Move subtest.docker_rpm() -> environment module, and switch to
using subprocess instead of autotest.client.utils.run for easier
addition of unittests.

Move subtest.known_failures() and SubSubtestCaller.is_known_failure() to
subtestbase for easier addition of unittests.
